Identifying Sec Lead
I'm concerned that reading the Guide is not enough there should be a step of delegation and or ownership that generates forward progress. Identifying a lead may instigate further action and implementation. I worry that a small office or a mid range at IG1 might say "well someone should do this" but then never assign it as a responsibility. I feel like this should be a first step.