minor formatting, links, and BP list updates
The old action list from an earlier draft was still sitting at the bottom of the page. I moved it in to Maturity 1. I think the intention was to remove those and have the 2 actions in maturity1, but I was hesitant to kill the content without confirming. If so, feel free to delete the rest of the list, @JTAGulous